% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridtools.R
\name{create.variable.name}
\alias{create.variable.name}
\title{Convert file name to variable name}
\usage{
create.variable.name(filename, prefix = NULL, fsep = .Platform$file.sep)
}
\arguments{
\item{filename}{character string}

\item{prefix}{character string: optional prefix to be added}

\item{fsep}{character used to separate path components}
}
\value{
character string: a variable name, see example below
}
\description{
Convert a file name into a variable name
}
\examples{
\dontrun{
create.variable.name("C:/my-path/my-file-name.Rd",prefix="res")
# Result: "res.my.file.name"
}
}
\keyword{utilities}
