% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_area.R
\docType{data}
\name{study_area}
\alias{study_area}
\title{Study area mask for the \code{landslides} dataset}
\format{
An \code{sf} polygon object representing the outlines of
the natural part of the RBSF study area.
}
\source{
Muenchow, J., Brenning, A., Richter, R. (2012): Geomorphic process rates of
landslides along a humidity gradient in the tropical Andes, Geomorphology
139-140, 271-284. DOI: 10.1016/j.geomorph.2011.10.029.

Stoyan, R. (2000): Aktivitaet, Ursachen und Klassifikation der Rutschungen
in San Francisco/Suedecuador. Unpublished diploma thesis, University of
Erlangen-Nuremberg, Germany.
}
\usage{
study_area
}
\description{
Study area polygon for the \code{landslides} dataset
}
\seealso{
\link{landslides}
}
\keyword{datasets}
