% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAGA-core.R
\name{rsaga.default.modules.path}
\alias{rsaga.default.modules.path}
\title{Internal functions that determine OS-specific default path in which modules might be located.}
\usage{
rsaga.default.modules.path(sysname = Sys.info()["sysname"], saga.path)
}
\arguments{
\item{saga.path}{character: path with SAGA GIS binaries, as determined (e.g.) by \code{rsaga.default.path}}
}
\description{
Internal functions that determine OS-specific default path in which modules might be located.
}
