% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAGA-core.R
\name{rsaga.get.usage}
\alias{rsaga.get.usage}
\title{Usage of SAGA command line modules}
\usage{
rsaga.get.usage(lib, module, env = rsaga.env(), show = TRUE)
}
\arguments{
\item{lib}{name of the SAGA library}

\item{module}{name or numeric identifier of SAGA module in library \code{lib}}

\item{env}{a SAGA geoprocessing environment as created by \code{\link[=rsaga.env]{rsaga.env()}}}

\item{show}{logical (default: \code{TRUE}); display usage in the R console?}
}
\value{
The character vector with usage information is invisibly returned.
}
\description{
\code{rsaga.get.usage} provides information on the usage of and arguments required by SAGA command line modules.
}
\details{
This function is intended to provide information required to use the
\code{\link[=rsaga.geoprocessor]{rsaga.geoprocessor()}} and for writing your own high-level interface
function for SAGA modules. R--SAGA interfaces already exist for some SAGA modules,
e.g. \code{\link[=rsaga.hillshade]{rsaga.hillshade()}}, \code{\link[=rsaga.local.morphometry]{rsaga.local.morphometry()}}, but there
are many more.
}
\examples{
\dontrun{
rsaga.get.usage("io_grid",1)
rsaga.get.usage("ta_preprocessor",2)
rsaga.get.usage("ta_morphometry",0)
# in SAGA GIS 2.1.0+, compare:
rsaga.html.help("io_grid",1)
# etc.
}
}
\seealso{
\code{\link[=rsaga.html.help]{rsaga.html.help()}}, \code{\link[=rsaga.geoprocessor]{rsaga.geoprocessor()}}, \code{\link[=rsaga.env]{rsaga.env()}}, \code{\link[=rsaga.get.modules]{rsaga.get.modules()}}
}
\keyword{interface}
\keyword{spatial}
