% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAGA-modules.R
\name{rsaga.target}
\alias{rsaga.target}
\title{Define target grid for interpolation}
\usage{
rsaga.target(
  target = c("user.defined", "target.grid", "header"),
  user.cellsize = 100,
  user.x.extent,
  user.y.extent,
  target.grid,
  header,
  env = rsaga.env()
)
}
\arguments{
\item{target}{character: method used for defining the target grid}

\item{user.cellsize}{Only for \code{target="user.defined"}: raster resolution
(in the grid's map units)}

\item{user.x.extent}{See \code{user.y.extent}}

\item{user.y.extent}{Only for \code{target="user.defined"}: numeric vectors
of length 2: minimum and maximum coordinates of grid cell center points}

\item{target.grid}{Only for \code{target="target.grid"}: character string
giving the name of a SAGA grid file that specifies the extent and
resolution of the target grid; this target grid file may be overwritten,
depending on the specifics of the SAGA GIS module used.}

\item{header}{Only for \code{target="header"}: list: ASCII grid header (as
returned e.g. by \code{\link[=read.ascii.grid.header]{read.ascii.grid.header()}}) or defined manually;
must at least have components \code{ncols}, \code{nrows}, \code{cellsize},
and either \code{x/yllcorner} or \code{x/yllcenter}.}

\item{env}{A SAGA geoprocessing environment, see \code{\link[=rsaga.env]{rsaga.env()}}.)}
}
\value{
a list containing information such as the bounding box of a target grid
}
\description{
Define the resolution and extent of a target grid for interpolation by SAGA
modules based on (1) user-provided x/y coordinates, (2) an existing SAGA grid
file, or (3) the header data of an ASCII grid. Intended to be used with
RSAGA's interpolation functions.
}
\note{
This function is to be used with RSAGA functions
\code{\link[=rsaga.inverse.distance]{rsaga.inverse.distance()}}, \code{\link[=rsaga.nearest.neighbour]{rsaga.nearest.neighbour()}}
and \code{\link[=rsaga.modified.quadratic.shephard]{rsaga.modified.quadratic.shephard()}}. Note that these are
currently only compatible with SAGA GIS 2.0.5 and higher.
}
\examples{
\dontrun{
# IDW interpolation of attribute "z" from the point shapefile
# 'points.shp' to a grid with the same extent and resolution
# as the (pre-existing) geology grid:
rsaga.inverse.distance("points", "dem", field = "z", maxdist = 1000,
    target = rsaga.target(target="target.grid",
    target.grid = "geology"))
}
}
\seealso{
\code{\link[=read.ascii.grid.header]{read.ascii.grid.header()}}
}
\keyword{interface}
\keyword{spatial}
